Fix rf/rz/r calculations to be more robust and accurate
Previously, we were mapping rz[⍸t=Z] to the enclosing z node, which was too aggressive. Instead, we want rz nodes to bottom out at the first Z node that they find, including self. Then, rf is always a strict indicator of an ancestor. Additionally, this helps to ensure that the r calculation is correct.