Parse ⍺ in all cases
In the cases where ⍺ is unbound, we should still parse it as if it were an array value. This is because the execution of that code is still technically required, and we want to signal an error right at the point at which the ⍺ is referenced in the resulting code.