Treat unparsed Z nodes as syntax errors
Now that we have introduced multiple types for operands, it is possible that there are some specializations of the source that cannot lead to a valid syntax tree. Previously, these would trigger a syntax error, however, now that we have multiple types, these syntax errors must be triggered at runtime and not at compile time.

The semantics we anticipate is that when a function is called with operands of a given type, the code execution will proceed (thereby ensuring that any state updates and effects are handled appropriately), until an unparsed Z node is encountered, at which point, these error nodes will trigger a syntax error.

This is the most faithful semantics that I can come up with at the moment, and it preserves observational equivalency as much as possible.

At the moment, I am deleting all of the nodes in a Z¯2 node because they are unnecessary, but conceivably, it might be better to do this in TT (conceptually), or not at all to provide space for managing these in some other way.