Create more favorable error reporting
You can only assign to ⍺ when it is a function-level, non-nested assignment. We can make sure that ⍺ is not bindable under other conditions. Additionally, because of this, we also want to ensure that all ⍺'s are treated as variables during the creation of bindings so that we can error when they appear in statements like ⍺ V←5.