Annotate the F kinds with information  about ⍺
We expand the range of potential types for our function nodes based on a new analysis.

We look for the type of the ⍺← binding that appears inside of a monadic function and annotate the F kind to indicate whether the monadic function call will bind ⍺ to a function or otherwise default to binding it to an array. This allows us to generate better code in GC.