Switch to using an explicit moper and doper type
We put in the code generation for the various calls that now leverage the fact that we have multiple type specialized function definitions. This requires both that we use moper and doper types that store the appropriate specialized function pointers, but also that we update how we handle monadic functions in the ibeams and elsewhere.

We take this opportunity to update the way that we generate code to reduce the amount of duplication we have and combine a few generators.