Disambiguate schizophrenic functions
Since we do not have the type information in the name of the function anymore, it is important to disambiguate the function names of the schizophrenic functions or they will clash with one another and trigger type errors.

This separates them out and creates two different underlying runtime function calls for them.