Release nested arrays in `release_host_data` instead of `release_array`
We don't want to always release the underlying nested values when we free an array. If the ref count for a value buffer is not zero, then we don't want to free those values. We only want to free those values when we are also freeing the host data buffer, otherwise we will invalidate our structures.