Don't let `squeeze_values` produce bad data
Previously the `squeeze_values` function was not doing what it should have. We need to ensure in all cases that we properly manage our releasing of previous data, and we also have to be sure that we don't overwrite data that we do not have sole ownership over.

If we are dealing with data that has a higher refcount than 1, we can't safely reuse that buffer to store newly squeezed data, so we need to allocate new data regardless in this case.

If we end up squeezing a nested array, we need to make sure that we don't delete all of the buffer data around it in cases where we aren't the sole owners and we also have to make sure that we properly manage the new buffer's ref count.

Previously we were acting under an assumption that we only had a single ref count, and we need to not do that.