Refine lexical scoping
This commit adjusts our treatment of lexical scoping. This is a deviation from the behavior in Dyalog APL. Specifically, we now treat local variables as we did before, but when encountering free variables, we allow for any forward reference to match, but we always match only once, to the nearest matching lexical contour.

This means that dfns with free references in them will only match a single lexical binding, and attempting to call a dfns before all of the forward references are bound will result in a VALUE ERROR. For example, consider this function:

      {X←5 ⋄ f←{g←{X+X} ⋄ Y←g⍬ ⋄ X←7 ⋄ Y+g⍬} ⋄ f⍬}

In Dyalog APL, this function will work because `g` will reference the `X←5` on the first call. In Co-dfns, we consider this an error, and make sure that `g` will only refer to the `X` appearing in the scope of `f`.