Convert ⎕NC calls to A1 nodes
In order to permit some of our code to be able to tell where it is within the overall structure, we allow for a very limited ⎕NC functionality, but requiring that we use a literal array that only looks up the values of formals that could potentially change, namely, ⍺, ⍺⍺, and ⍵⍵.

We do this by transforming these ⎕NC calls into literal arrays (evaluating them at compile time) using the kind of the referent function, so that we do not need to do any computation at compile time to get this information.

A problem with this is that we still need to do something dynamic at runtime to handle being in different situations. I think that this will be fixed up easily enough if we have appropriate constant propagation and inlining and dead code elimination and such.