Make all free variables mutable
One of the previous issues prior to this patch with forward references is that a non-mutable bindings will be captured at the point of the dfn's closure creation, but subsequent mutations of that cell will not propagate down to the free variable reference in the body of the dfn. The prior semantics was to have only the value as it stood at the time of the dfn definition be used throughout the dfn. However, the semantics of dfns expects the value to change with mutations to the name.

This patch allows for mutations in the name by treating all free variable references as mutable. This is overly cautious, because this only matters if you have a forward reference to the free variable. If there are only back references to the variable, then it is safe to treat the variable as immutable. However, we do not currently analyze the code to determine if this is the case, so the safe thing to do is to always treat them as mutable. This transparently handles any forward references that we may have.