Improve marking of mutable variables
We want to mark the E4 target as mutable, but not anything in the E6 node. Important: even though we could get away with not marking E4 targets as mutable right now if they are not also part of a free reference somewhere, this would break in the future if we introduced something like constant propagation, and we don't want that to happen. We are better off just treating these as mutable right now. Maybe later we can optimize this.