Mark sure to release the original value correctly
When we have an E4 node that is just a modified assignment, we want to make sure that we release the original value twice, because there will be two references to it, the stack reference and the binding reference. However, if we are in a bracket assignment, there will only be a single reference. We still need to do the release, but we also need that data, and we don't want to free the array after the mutation. Instead, we will rely on our get and set functions creating an extra hold on the data after assignment, so that the release will not cause any harm.