Improve some reuse calculations and always disable reuse for nested arrays
We want to ensure that we are always releasing the nested elements when we squeeze them. In order to keep us in a consistent state, it is best to ensure that we always do a copy when extracting values from a nested array structure, rather than doing an in place update.