Mark axis operator specially
We mark the Axis operators specially because we want to do special things in the code generator for managing their closures and handling variant options and the like. This means we want a different node type for them. We also do not need the [ primitive entering into the stack because of this, so we will remove that from the tree before it gets lifted up.