Make func->fv an independent array
We want to enable some functions to actually point to other functions for their free variables. To do this, we need to divorce the fv[] array from the allocation itself. We do this by have an independent fv void ** field that we then point to the internal fv_[] structure by default. We can then point fv somewhere else if we want it to use another free variable reference from somewhere else.