Define derive_func_opts
The derive_func_opts() function is intended to derive a new function that is meant to hold a number of variant options. The number of variant options are given by `fs` and we derive a new function that shares the free variables of the original function. We need to make sure that we create a reference to the original function in our frame, which we do by setting fv[0]. This way, releasing the derived function will appropriately clean everything up if it happens that we are the last reference to the original function.