Refactor the ⍺← deletion pass
This is a cleaner and probably faster way to do this. We lift some of the conversion passes up so that they do not need to work on as much data and so that the B nodes are in a nicer format for us to work on the ⍺← recognition code.