Improve the reliability of local variables
When we have an ⍺← deletion occur, it is possible that there was a binding in that expression that would be removed. If that binding is needed somewhere else in the code, then we want to signal a runtime VALUE ERROR. In order to do this, we still need to recognize that a local binding should exist, even though we have deleted it. To do this, we adjust our `lv` calculation so that we will infer the need for a local variable binding if there exists a local variable even though there may not be a local binding.