Make the body of ⍠← bindings execute conditionally
Just like ⍺←, it is quite useful to be able to know whether a given option has been bound or not. The only way to do this without needing to create additional functionality to query the value of the option directly is to make the body of ⍠← conditionally execute so that we can put a flag inside to set it when it has been executed.