Fix up the handling of VEACH and get TEST 9 passing
This is a pretty large diff, but is relatively simple. We needed to get the rest of TEST 9 working. In order to do this, there was an issue in VEACH handling scalar values.

The main issue is that when working over a scalar value, we had a single array that we just mutated to slide along the input array. This was an issue if a reference to this array persisted in the resulting expression. Doing so could result in us mutating the intermediate results of the VEACH computation, leading to all sorts of bad things.

Now, we explicitly make a new array for each call to ⍺⍺ in VEACH that we make, and we release those after the fact, so that references to the array can persist after the fact. This preserves the appropriate references.

While doing so, we needed to greatly refactor this to make it much easier to handle. To do this, we introduce two much cleaner switches over the whole range of data types for arrays, `MONADIC_TYPE_SWITCH` and `DYADIC_TYPE_SWITCH`, which we also simplify to only have a single expression that needs to be generic over all the types.

These macros are much easier to use and more consistent, as well as easier to recall. They should replace our other switches over array types, but we are not doing that in this diff.

We would be able to get by with only a single macro to case over all the types, but doing so would require that we have support in C for recursive macro expansion, which we don't. This means that we need a switch for the monadic and the dyadic cases explicitly. Yuck.

To make this all work better, we have makers for all of the data types, and not just the typical simple types. This ensures that we handle retention appropriately as well as ensuring that we can make the code have the same general shape for all types. This means that we change the `mk_scalar_` class of functions to `mk_array_` functions, since now we may not always be "making" a scalar.

Finally, we were replicating some important logic that needed to be extracted out, which is the `array_get_host_buffer` function. This logic is needed whenever we need to get a host buffer of the data, but where we don't actually want to work over a full array, just iterate or maneuver with the buffer of values. We need to have a flag to indicate whether it was necessary to allocate a new buffer for DEVICE allocated data, or whether we can just return a pointer to already host allocated data.

The end result is that VEACH is now much more reliable and simpler, and TEST 9 is now passing.