Refactor squeeze; fix bool issue with macros
We had an issue where bool will expand itself out into another type when we have more than a single macro expansion pass happening, and we don't want that, so we get around it by using a suffix that isn't going to clash with existing defines.

This allows for refactoring the squeeze code to be much cleaner.