Disable axis operator checking in parser
This is a little dangerous right now, but we want to allow our code to generate even in cases when we are applying an axis operator to a function that we don't know yet. This means we will need to have a way to verify at runtime that we have axis operator support.