Calculate free variables on the basis of C and not the body of F nodes
It can happen that after parsing we may have removed some references to particular variables or the like, deleted code, or otherwise caused a situation in which one specialization of a function does not adequately match another in terms of free variables.

This can cause a miscalculation of free variables and mess with code generation.

In order to maximize the overall independence of specialized functions, as well as allow us to do this free variable calculation late rather than early, we calculate free variables on the basis of the contents of C nodes, rather than looking at the bodies of the specialized functions, as this represents a single source of truth for all specialized functions.