Account for schizophrenic functions with bracket indexing on the left
At this point we have rationalized the F[X] Axis operator, but the V[...] case still has a ¯1.0 node sitting around. We know that this will eventually become an A¯1 node, but by this point, we don't do that because we want to be able to handle the cases where V[...] is an assignment target. Thus, we should recognize here that a ¯1.0 node on the left of a schizophrenic primitive means that this primitive has an array left argument and parse them accordingly.