More explicit and particular in our casting
Before this, we were assuming that we would apply our real operations to the inputs at their source type, and then cast to the result type after the operation was complete. This works sometimes, but doesn't work when we are at the edges, because it doesn't account adequately for the overflow and underflow cases.

To account for these edge cases, we need to cast all inputs to the intended output type first, before we do the computation. It is also important to note that this only works in those cases where we know that the output type is at least as large as the largest input type. For example, this will not work when we are dealing with comparison types, which always move things down into their Boolean types, and so we require a different set of macros for handling those cases.