Begin implementing fast paths for reduction
Now is the time to start identifying and improving the performance of the APL algorithms.

The first target is reduction. With specific operands, we know that there are specific runtime codes that we can use which will be much faster than the general ones that we are using here.

We also want to assume that we will speed up our handling of scalar codes at some point in the future, which should make the performance of the general case go faster as well, but for now we are going to focus on the specific fast paths that are necessary for getting other things to go fast.

Specifically, this means that we need to focus on getting indexing to go faster, which means focusing on those basic primitives that need to be very fast when doing indexing operations. This includes reductions. This is the first step towards that.