Force `spn` to stay constant
If we reference `spn` inside of our code instead of leaving it as a primitive constant, then we can run into trouble with the way our analysis of free variables at the moment forces `spn` to become a boxed value at that point. The easiest way around this is to inline the reference to `spn` if we want to do this. In the future we may want to look at fixing this in a more general way.