Implement faster bracket indexing
We have special cased the bracket assignment now and are doing the same thing as we did for indexing. This removes our use of `flattend_idx` from the code, but there is still work to be done to ensure that ↑ works in the cases where it was depending on the use of `flattend_idx`.