Statically allocate at least 4 dimensions for each array
We are working to reduce the allocation costs that come with these arrays. The way that we want to do this is to try to stack allocate the array headers as much as possible. To do this, we need to make arrays have a reliable static size, instead of relying on them having a dynamic size. The first step in doing this is to make sure that they have enough space to statically store most of their shape information for most arrays.

We do all of this by statically allocating at least 4 dimensions for each array. The plan is that we can use this extra space to do additional work if we want. For instance, if we have scalars, we can embed the values right into that memory space. If we need to have larger ranks, then we can just allocate those if we want. This should allow us to make many things more static.