Switch to the transpose method of axis operators for scalars; implement encode without appeal to axis operator
This also requires that we rework the way that we handle encode so that we do not need to use the axis operator. This does have the advantage of making encode a little faster in some cases and also able to scale to arbitrarily large number systems instead of being limited by our representational power.