This code doesn't quite work, but I want to keep it here because I want a record of this code and what I tried
The rational_val calculation is an IEEE binary rational approach, but this is too accurate and doesn't actually give us answers that come close to what Dyalog uses.

The rational ibeam on the other hand comes very close, but for some reason it is slightly different on a few values, which I haven't figured out why. Doing this same code in APL seems to work fine, and is what I've ended up needing to use.