Revert keyword splitting
The function of keywords is too intimately tied up with the semantics of lines to make it possible to do splitting of keywords early in the parser. Furthermore, the semantics of keywords is tied up with lines enough that there is a much stronger general association between keywords and lines than appeared at first. This makes it more likely that another approach should be used, such as marking a line directly instead of having K and Z nodes live at the same level.