Add support for printing ⎕← and ⍞← to stdout
This is work to enable printing for basic executables (not for DWA).

Unfortunately, most terminals are not 16-bit enabled, which removes our ability to use the 16-bit representations for characters that we were using when printing or working with external interfaces (largely driven by Windows and the Dyalog DWA IME).

Instead, we need to switch to using UTF-8 as our primary source input and output format. Internally, we still use the codepoints to represent characters, however, when printing, we print them out as UTF-8 strings using our new printing functions. This also means that we need to alter how we handle error reporting and tracing in order to make sure that we are printing out those strings in the same format.

To account for this, we need to change our tracing interfaces to all use plain char strings that contain UTF-8 data and also change our debugging interface to ArrayFire. Fortunately, this is more in line with how ArrayFire does things, so it's not so bad.

This also means we need to change how we interface with Dyalog. This happens in two ways. When we send errors back through DWA, we pass them as 8-bit UTF-8 character arrays, which means that we need to decode them when they come in before we print them. Additionally, if we use Exec, we need to decode the returned data from ⎕CMD first, including line ending conversions, and then we can print that data as well.

This seems to fix up the printing behavior that we have, and works a little more reliably and feels less brittle than using the 16-bit stuff, so I'm happy with this overall, but it's a big change.