Fix an issue with mtype coming out of synch with the array values
In the previous code, it was possible that one value of the two `tgt` and `r` might be nested coming in, but might squeeze down to a smaller representation. We calculated `mtype` before we called the array promotion code, which meant that the `mtype` value was based on the pre-squeezed values. This could result in calling `cast_values` on an array that was promoted to device storage assuming squeezed arrays with an `ARR_NESTED` type, resulting in an error.

To fix this, we move to squeezing the arrays and doing promotion before we do the calculation of the `mtype`. This may cause us to migrate the `r` value even if we might not really want to on a global stage, but we can deal with that later.