Fix an issue with correctly identifying terminating bindings
We want to mark expressions that are the last expression in a dfns as a return statement even in the case where they are binding values. However, the previous code assumed that all expressions were accounted for by p[j], but j does not include G nodes, so it would be possible for you to return early if the final expressions in a dfn were all guards.