Use the init functions for main and don't make an explicit main()
I don't like that we have to have a special function to use for our main function. Instead, we know that we already have initialization happening inside of a namespace, so why not use that as the main function? This helps to reduce the total name clutter, simplifies the generated code and also the mental model of the whole thing. It also has the benefit that multiple namespaces can be used as a main function simultaneously and easily allows us to distinguish between functions that we may want to do stuff in main and those that we don't.

The biggest disadvantage here is that some things that work in Dyalog APL for functions do not work when you have them inside of a namespace directly. I think this is a mistake, and I think that we should be able to do both, so I'd rather do the "right" thing here instead of sticking with the limitations inherent in Dyalog for something that Dyalog isn't even really doing right now.

This also creates a few more assumptions. One, that we are going to use ⎕SIGNAL  for our abnormal exits, rather than a return code, which I think is fine by me. It also assumes that we are going to need some way to pass in the command line arguments to the namespace if we ever want them. This does mean we introduce a little more complexity there, but I think the flexibility is worth it.