Handle complex values correctly on the device
The behavior of af_min_all and af_max_all is not what I thought, and they are actually computing a specific min or max values even for complex values. This means that I need to do something a little different, and first determine whether we have complex values in the first place, and exit out early in that case. Otherwise, I can continue under the assumption that I only have real values, and then I don't have to worry about things.