Add core transformations to support strand assignment
We divide these strand operations into three main pieces.

The first piece remaps the binding target to use S nodes instead of A and V nodes for stranded values. This allows us to treat them specially for stack operations.

The second operations is to ensure that the S nodes are annotated with their child counts.

The third operation lifts the binding target into a flat structure ahead of normal expression flattening to ensure that they retain their original order, thus executing them in a pre-order traversal pattern instead of in the RPN style that we use for the rest of the expressions.

Finally, we remove the B0 node as superfluous after we are done with these operations.