Eliminate boxes
In order to reduce allocation overheads, we now don't use dynamically allocated boxes. This limits us at the moment in terms of closures escaping their contexts, but we can deal with that later.

To make this change, we make every boxed value (free variables) be passed by ** in the closures, instead of inside of a box. This is sufficient as long as the closures do not escape. This requires some bookkeeping to make sure that we don't release these boxes, and that we reference them appropriately.