Eliminate unnecessary grading calls through the AF libraries
There is a major bottleneck in performance that this fixes. We now do not need to use the device code to handle the grading functions, which we also use in the Axis operator, so this is a critical performance issues across most of our code.

This implements a simple C based version that should avoid the need to go to the device.