Use static space in structure to avoid alloc_array malloc
We use the static space that we've created in the cell_array struct to store data if there is room, which means we do not need to malloc space for most of our arrays when we alloc_array() space for them.

We do copy data from one static allocation to another rather than sharing buffers. This is to ensure that a static buffer is never shared.

It also doesn't make much sense to have all of the squeeze code in its own file since there is so much overlap, so we're putting it all into array.c.

To help avoid more trouble, we've also moved most of the logic around handling vrefc into the array.c file and out of the other files, like ibeam.c, so that we can control that a little better. This introduces the array_share_values function.