Make handling of stranding and brackets more reliable
This makes an adjustment to the way that we parse strands and brackets to help improve reliability a bit. The previous problem was that if you had something like ⍺⍺[...] ⍵, where ⍺⍺ was a value, then the stranding would not recognize the whole thing like (⍺⍺[...])⍵, and this would result in bad parses.

To fix this, we look for such expressions in advance and wrap them as Z nodes, and then allow stranding to take place. Afterwards, we do the standard wrapping for V[X;...] nodes, which should happen after we have wrapped stranded values on the left into a single A7 node.