Fix up the wrapping handling for A¯1 nodes
Our previous wrapping code doesn't work with the new wrapping that we have done. We need to ensure that we have no extra Z node between the siblings and the A¯1 node by the time we get to this point, as we need to know about A¯1 nodes in order to parse things like bracket assignment correctly.