Add an error/syntax error check for E6 nodes that contain function values
Since we are switching away from the stack model, we need to be a little more careful on our types here. This will make sure that we don't allow for generating code/nodes where we are trying to bind a function value as an array in an indexing expression such as A[⍺⍺] where ⍺⍺ is a function.