WIP: Parsing code for type inference
This is the beginning of some work to work on getting type inference to work in the presence of trad-fns strand assignment even in dfns and elsewhere. This should unify the stranding assignment behavior between the two systems and also allow us to handle modified assignment in the cases of named modified assignment.