WIP: More parsing notes
This is a sketch under the assumption that we will be treating parsing of binding and E4 nodes as a secondary pass over already parsed ← primitives which are initially treated as schizophrenic functions.

We do this after getting rid of Z nodes because that should mean that we don't have as many node types to deal with or as deep a tree.

We want B nodes to be constructed any time that we may have anything that is bound at all, which means that we may have internally nested bracket and selective assignments inside, but those will fall out okay over time.

We can probably just assume things are fine here and delay checking until a separate pass for bad binding targets.