WIP: Accept the present dfns behavior and land on a semantics of inference of trad-fns
In the inference pass, we can assume that we can propagate types all the time. For dfns, we expect binding values to point to binding names which can then be safely propagated through always for dfns, and for trad-fns, we expect that the first binding node will point to the header binding, which will in term be the binding to which all other variables refer, thus creating the appropriate propagation loop.