Reintroduce lx, xn, and xt fields; fix up closure creation
It just doesn't make sense to put these into some other phase when really we only care about the PS output.

Adjust the closure creation to make sure that we create closures around the dynamically scoped trad-fns references as well as try to ensure that our free variables are created at the same time. This means we need to delay closure creation until we know what those trad-fns references are.