Base exports off of the C value for trad-fns
We have the name stored in the C node that contains trad-fns. We can use this as a binding form and we should treat this as such, since specialization may result in many T nodes for a single trad-fns, and there is otherwise no binding node for trad-fns.

Alternatively, we could make a B node to enclose the C node of trad-fns, but I think it might make sense to keep the C node like this, so we'll leave it for now and see how it goes.