Add preliminary support for parsing with globally free values
This is a somewhat subtle addition.

When we have variables that bind global names that are not bound at the top level explicitly (such as a trad-fn binding a free variable that is not shadowed anywhere else), we have globally free values that we need to type somehow.

To do this, we first create links to a newly created global H containing all these global free variables, and then we try to use any convenient first binding to accomplish our ends.

There are some pathological cases where this won't work, but we're going to see how often those show up in the real world first.

We also take the somewhat controversial step of making these exported values as well. I don't know if this is the right call, and we might want to remove these from the export set at some point.