Pass through complex axis operator statements
If we have an Axis operator form of the type F[A;B;...] then we will just consider this the same as an Axis operator call of the form F[(A)(B)...]. This helps to handle cases where we have an operand ⍺⍺ or ⍵⍵ that we index using ⍺⍺[A;B;...] forms.

Since the Axis operator should already be robust against nested values in most cases, this should have little to no real effect on the runtime reporting of the code, but it does simplify AST generation and make things a little easier for static analysis. In practice, many such cases will still be mapped to Z¯2 nodes.