Relocate handling of terminal bindings
We have to be careful where and how we treat these terminal bindings as it relates to tracking free variables, so we do this extra binding linkage after we have already considered free variable bindings in the trad-fns. In the case of dfns it won't matter, since these forms overlap, but with trad-fns we need to make sure that we still have an authentic vision on the free variables, even if we treat some assignments as having known binding types internally instead of linking against their original 1st assignment or binding location.