Fix parsing of ∘. as the first item in a group.
∘. is a special form, unlike any other primitive or user defined
function (be it a dop or tradop) it is a monadic operator that only
accepts one operand on the _right_.  Monadic operators normally
accept their operands only on the _left_; under the assumption that all
monadic operators behave in this way the next line propagates the
kind of the first child of a group to the parent: if the first child
has kind 3 (i.e. is a monadic operator) this is correct assuming
the entire group is well-typed.  Since ∘. breaks this assumption we
have to handle it separately.