Adjust r values to account for nested namespaces and closures around all F nodes
We need to support nested closures, which means that we now have closures which are direct parents (not separated by Z nodes) in the tree. We need to handle this. We also need to account for the fact that we now wrap a C node around every F node, even F0 nodes.

The extra checking granted by the ¯1 modification isn't particularly helpful at this point, so we remove it.