Treat global variable references specially
We now have essentially three classes of references. Local references are positionally aware (lexical ordering), while free references are only sensitive to lexical contours, thus enabling forward references, and they exist in C-sets, but global references cannot populate C-sets, but instead, work off of H-sets in F0 nodes, but ignore lexical positioning.

This deals with the case where we have an unbound variable reference in an F0 body that occurs before a binding to that name. In such a case, we do not treat it as free, but instead, still resolve it as linking to the same binding as the latter assignment. This unifies the links, but finding the unbound references will be a little harder.

We could leave such names unbound, but doing so would make parsing more difficult and would also mean that all vb=¯1 valued V nodes would no longer exist solely in H-sets by the time we were done with variable resolution.