Link Z nodes and bindings
We are doing this here instead of doing this up where we did it before because we want to do a little more than we did up there and we also don't need to use this information until we get here, so we might as well delay processing.

We also are explicit about the handling of the bindings because we may have different binding forms to handle, so we want to be specific about the assignment forms that we want to use here, in this case just the ← and ∘← cases.