Adapt type inference to closure-based resolution and namespace references
So, this patch is a little subtle.

Firstly, it adapts the code to account for the fact that we have closure-based resolution now. This has to deal with the fact that vb is different, but even more important is the fact that we no longer have a canonical binding anymore. Instead, bindings are in H nodes and some ⍺← forms. These all have vb=¯1. They do not point to any canonical binding anymore.

This presents a challenge for type inference, as we need to figure out how to infer the types without a canonical binding. The advantage is that we no longer need to try to figure out what the first binding of a form is, which is generally impossible in the worst cases.

We make use of a very important fact: any binding that is ever set to a function cannot ever be set to an array first. There is no non-error state in which a binding is set to an array first and then to a function. This means that we either have a nameclass error or modified assignments in other cases where we don't bind function types to the binding.

We also know that binding forms for function (including mop and dop) types are much more restricted, and easy to track in the z-target vbs. Knowing this, we can do all the inference we need by simply looking for every place where we have a function binding, and infer only function types. If we propagate all function types correctly, then we can assume everything else is an array type after we are done. Since we are safe in opportunistically setting function types (this is not true for array types), this works out.

A situation arises in this type inference, though, where we can stall out in cases where we have dops with array-typed strand right operands. In such a case, we have to do all the inference we can first, and then build a dependency graph of the dop expressions to find the leaves. The leaves we know must then be only array valued right operands, and we can safely treat such expressions as ending in a function typed value. We can then continue on, each time handling stalls by inferring the types of the leaves of the dop expression dependency graph built at each stage (we cannot reuse the dependency graph directly because nodes will be removed and become new leaves later).

Furthermore, the presence of namespace references throws a serious wrench in all of this, because it is impossible to tell the difference between singleton expressions and compound expressions in the presence of dots that are actually namespace references. However, we know that in order to handle namespace references generally, we must have some way to resolve namespace references before type-inference, because inference must infer off of the appropriate binding inside of the correct namespace. This implies that we can tell what dots are namespace references and what dots are inner products *before* we get to the inference stage. Thus, we make the assumption that we will be parsing namespace references into N nodes before we do inference. This is not yet done, but we have written this inference code under this assumption.

While a deceptively small amount of code, the implications here are large, so I hope I haven't forgotten any important bits in the notes!